Move entity specs into own file
